requirements.txt
streamlit>=1.40.0
playwright>=1.49.0
google-genai>=1.0.0
nest-asyncio>=1.6.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.8.0
matplotlib>=3.8.0
plotly>=5.18.0
Pillow>=10.0.0
python-dotenv>=1.0.0
